move new relic callbacks into UXCapture.jsx and remove UXCaptureNewRelicConfig component
